refactor: 단위 테스트 패키지 위치 수정 (#12)
