+++ src/test/java/com/example/green/dummy/ApiDummyController.java
@@ -1,4 +1,4 @@
-package com.example.green.unit.dummy;
+package com.example.green.dummy;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
