+++ src/test/java/com/example/green/global/HealthControllerTest.java
@@ -1,4 +1,4 @@
-package com.example.green.unit.global;
+package com.example.green.global;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import com.example.green.global.HealthController;
 
 public class HealthControllerTest {
